djrssproj/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
bs4==0.0.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
cron-descriptor==1.4.0
distlib==0.3.8
distro==1.8.0
Django==4.2.8
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-timezone-field==6.1.0
djangorestframework==3.14.0
dnspython==2.4.2
docopt==0.6.2
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
eventlet==0.34.3
exceptiongroup==1.2.0
feedparser==6.0.10
filelock==3.13.1
flower==2.0.1
frozenlist==1.4.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
humanize==4.9.0
idna==3.6
importlib-metadata==7.0.0
kombu==5.3.4
loader==2017.9.11
Logbook==1.7.0.post0
multidict==6.0.4
openai==1.5.0
packaging==23.2
pendulum==3.0.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg==3.1.15
psycopg-binary==3.1.15
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
termcc==2019.8.2
terminaltables==3.1.10
tiktoken==0.5.1
time-machine==2.13.0
tldextract==5.1.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uWSGI==2.0.23
vine==5.1.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.12
worker==2018.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
